backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
sqlmodel>=0.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
pytest>=7.0.0
httpx>=0.25.0